Decouple MoveResolver traits from Storage traits
In some of the places where we want to use the layout resolver, we
don't have access to all the information to implement `StorageView`,
but we only need to use the `BackingPackageStore`.

To support these cases, we need to provide a state view to
`LinkageView` that can serve packages, and then shim the Module and
Resource resolution functions.

This requires that `LinkageView` no longer accept a reference
parameter, but a value (so we can inject state into it) -- similar to
the trick we pulled to allow `Session` to own a `LinkageView`.  Then
implementors (like `TemporaryStore`) that only need to be passed by
reference can be passed by explicit reference,

i.e. `LinkageView<'state, TemporaryStore>` becomes
`LinkageView<&'state TemporaryStore>`.

However, we cannot impose a `&'state S: StorageView` bound, because
some APIs in `StorageView` require a mutable reference, so we needed
to decouple the resolution traits (now `SuiResolver`) and the storage
traits, to impose twin constraints:

```
fn ...<'state, S>
where
  S: StorageView
  &'state S: SuiResolver
```